requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-countries==7.5.1
django-filter==23.4
djangorestframework==3.14.0
Markdown==3.5.1
Pillow==10.1.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
